Construct route map at compile-time (#12)
Don't want to spend any runtime on constructing the hashmap.

I observe no meaningful performance difference, which is probably
because constructing a (currently) tiny HashMap just doesn't take very
long at all, so it's fine to do it for every lambda invocation. That,
combined with the fact that the actual lookup of phf is on the slow side
at the moment (ref rust-phf/rust-phf#349 + rust-phf/rust-phf#88), means
that it's probably expected for this to not be much faster. However, it
should become faster as the size of the HashMap grows (i.e., there are
more short-links), and should also be faster as phf improves. So, let's
land nonetheless.